Extract Express.js utils
