+++ eslint-bridge/src/rules/utils.ts
@@ -582,3 +582,76 @@ export function flatMap<A, B>(xs: A[], f: (e: A) => B[]): B[] {
   }
   return acc;
 }
+
+export function isMethodInvocation(
+  callExpression: estree.CallExpression,
+  objectIdentifierName: string,
+  methodName: string,
+  minArgs: number,
+): boolean {
+  return (
+    callExpression.callee.type === 'MemberExpression' &&
+    isIdentifier(callExpression.callee.object, objectIdentifierName) &&
+    isIdentifier(callExpression.callee.property, methodName) &&
+    callExpression.callee.property.type === 'Identifier' &&
+    callExpression.arguments.length >= minArgs
+  );
+}
+
+export namespace Express {
+  const EXPRESS = 'express';
+
+  /**
+   * Checks whether the declaration looks somewhat like `<id> = express()`
+   * and returns `<id>` if it matches.
+   */
+  export function attemptFindAppInstantiation(
+    varDecl: estree.VariableDeclarator,
+    context: Rule.RuleContext,
+  ): estree.Identifier | undefined {
+    const rhs = varDecl.init;
+    if (rhs && rhs.type === 'CallExpression') {
+      const { callee } = rhs as estree.CallExpression;
+      if (getModuleNameOfNode(context, callee)?.value === EXPRESS) {
+        const pattern = varDecl.id;
+        return pattern.type === 'Identifier' ? pattern : undefined;
+      }
+    }
+    return undefined;
+  }
+
+  /**
+   * Checks whether the expression looks somewhat like `app.use(m1, [m2, m3], ..., mN)`,
+   * where one of `mK`-nodes satisfies the given predicate.
+   */
+  export function isUsingMiddleware(
+    context: Rule.RuleContext,
+    callExpression: estree.CallExpression,
+    app: estree.Identifier,
+    middlewareNodePredicate: (n: estree.Node) => boolean,
+  ): boolean {
+    if (isMethodInvocation(callExpression, app.name, 'use', 1)) {
+      const flattenedArgs = flattenArgs(context, callExpression.arguments);
+      return Boolean(flattenedArgs.find(middlewareNodePredicate));
+    }
+    return false;
+  }
+
+  /**
+   * Checks whether a node looks somewhat like `require('m')()` for
+   * some middleware `m` from the list of middlewares.
+   */
+  export function isMiddlewareInstance(
+    context: Rule.RuleContext,
+    middlewares: string[],
+    n: estree.Node,
+  ): boolean {
+    if (n.type === 'CallExpression') {
+      const usedMiddleware = getModuleNameOfNode(context, n.callee)?.value;
+      if (usedMiddleware) {
+        return middlewares.includes(String(usedMiddleware));
+      }
+    }
+    return false;
+  }
+}
